fix: update all use paths after web/channel_adapter move
Update crate::web -> crate::channels::web and
crate::channel_adapter -> crate::channels::adapter across all files.
